refactor(structure): relocate non-type files from types/ directory
- Move awardIcons.tsx from types/award/ to components/award/
- Move PilotConstants.ts from types/pilot/ to constants/
- Move skillCatalog.ts from types/campaign/skills/ to constants/campaign/
- Move pilot-mech-card.ts from types/ to types/pilot/
- Update all import paths across codebase
- types/ directory now contains only type definitions

Ultraworked with Sisyphus
