+++ src/components/award/AwardBadge.tsx
@@ -11,7 +11,7 @@ import {
   getRarityColor,
   getRarityBackground,
 } from '@/types/award';
-import { getAwardIcon, getRarityStrokeWidth } from '@/types/award/awardIcons';
+import { getAwardIcon, getRarityStrokeWidth } from './awardIcons';
 
 // =============================================================================
 // Types
