+++ src/types/pilot/pilot-mech-card.ts
